Added frontend UI

+++ backend/controllers/listController.js
@@ -1,9 +1,10 @@
 const asyncHandler = require('express-async-handler')
 
-const List = require('../models/listmodel')
+const List = require('../models/listModel')
+const User = require('../models/userModel')
 
 const getList = asyncHandler(async (req, res) => {
-    const list = await List.find()
+    const list = await List.find({user: req.user.id})
     res.status(200).json(list)
 })
 
@@ -14,7 +15,8 @@ const setList = asyncHandler(async (req, res) => {
    }
 
    const list = await List.create({
-    text:req.body.text,
+    text: req.body.text,
+    user: req.user.id,
    })
     res.status(200).json(list)
 })
@@ -25,6 +27,18 @@ const updateList = asyncHandler(async (req, res) => {
     if (!list){
         res.status(400)
         throw new Error('List not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if(global.user.toString() !== user.id){
+     res.status(401)
+     throw new Error('User not authorized')   
     }
 
     const updatedList = await List.findByIdAndUpdate(req.params.id,req.body,{
@@ -42,8 +56,20 @@ const deleteList = asyncHandler(async (req, res) => {
         throw new Error('List not found')
     }
 
+    const user = await User.findById(req.user.id)
+
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if(global.user.toString() !== user.id){
+     res.status(401)
+     throw new Error('User not authorized')   
+    }
     await list.deleteOne()
-   res.status(200).json({ id : req.params.id})
+
+    res.status(200).json({ id : req.params.id})
 })
 
 module.exports = {
